+++ tests/Sniffs/Spacing/SpacingTestSuite.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace XynhaCS\Tests\Sniffs\Spacing;
+
+use PHPUnit\Framework\TestSuite;
+
+final class SpacingTestSuite
+{
+    public function getTestSuite()
+    {
+        $suite = new TestSuite('XynhaCS_Spacing');
+
+        $suite->addTestSuite(PropertySpacingTest::class);
+
+        return $suite;
+    }
+}